Added validation for POST requests

+++ API/Models/Subject.cs
@@ -6,7 +6,7 @@ public class Subject : BaseClass
 {
   public string Code { get; set; } = string.Empty;
   public string Title { get; set; } = string.Empty;
-  public short ECTS { get; set; }
+  public byte ECTS { get; set; }
   public bool IsObligatory { get; set; }
   public FieldOfStudy FieldOfStudy { get; set; }
   public Professor? Professor { get; set; }
